elasticsearch dependency update and refactor to use validate_study method in ws/validation.py
