+++ app/ws/MapStudies.py
@@ -33,6 +33,7 @@ from flask_restful_swagger import swagger
 from owlready2 import urllib
 
 from app.ws.mtblsWSclient import WsClient
+from app.ws.study.study_service import StudyService
 from app.ws.utils import log_request
 
 wsc = WsClient()
@@ -42,12 +43,11 @@ class getStudyInfo():
 
     def __init__(self, studyID, user_token):
         try:
-            url = 'https://www.ebi.ac.uk/metabolights/webservice/study/' + studyID
-            request = urllib.request.Request(url)
-            request.add_header('user_token', user_token)
-            response = urllib.request.urlopen(request)
-            content = response.read().decode('utf-8')
-            self.study_content = json.loads(content)
+            m_study = StudyService.get_instance().get_study_from_db_and_folder(studyID, user_token,
+                                                                               optimize_for_es_indexing=False,
+                                                                               revalidate_study=True,
+                                                                          include_maf_files=False)
+            self.study_content = {"content": m_study.dict()}
         except:
             print('cant find study', studyID)
 
@@ -86,14 +86,8 @@ class getStudyInfo():
 
 
 def searchStudies(query, user_token, feature='factor'):
-    # list of all studies
-    url = 'https://www.ebi.ac.uk/metabolights/webservice/study/list'
-    request = urllib.request.Request(url)
-    request.add_header('user_token', user_token)
-    response = urllib.request.urlopen(request)
-    content = response.read().decode('utf-8')
-    j_content = json.loads(content)
-
+    study_id_list = StudyService.get_instance().get_all_authorized_study_ids(user_token)
+    j_content = {'content': study_id_list}
     import re
 
     def atoi(text):
