+++ app/ws/assay_protocol.py
@@ -18,15 +18,14 @@
 
 from flask import abort, jsonify
 from flask_restful import Resource
+from app.ws.study import commons
 from app.ws.isaApiClient import IsaApiClient
-from app.ws.mtblsWSclient import WsClient
 from app.ws.utils import *
 from flask_restful_swagger import swagger
 import logging
 
 logger = logging.getLogger('wslog')
 iac = IsaApiClient()
-wsc = WsClient()
 
 
 class GetProtocolForAssays(Resource):
@@ -86,7 +85,7 @@ class GetProtocolForAssays(Resource):
         logger.info('Assay Table mapping: Getting ISA-JSON Study %s', study_id)
         # check for access rights
         is_curator, read_access, write_access, obfuscation_code, study_location, release_date, submission_date, study_status = \
-            wsc.get_permissions(study_id, user_token)
+            commons.get_permissions(study_id, user_token)
         if not read_access:
             abort(403)
