+++ app/ws/db/models.py
@@ -7,32 +7,33 @@ from app.ws.db.utils import datetime_to_int
 
 
 class UserModel(BaseModel):
-    userId: int = Field(..., alias="id")
-    email: str = Field(..., alias="email")
-    userName: str = Field(..., alias="username")
-    dbPassword: str = Field(..., alias="password")
-    userVerifyDbPassword: str = None
-    joinDate: datetime.datetime = Field(None, alias="joindate")
+    address: str = Field(None, alias="address")                         # excluded from es
+    affiliation: str = Field(None, alias="affiliation")                 # excluded from es
+    affiliationUrl: str = Field(None, alias="affiliationurl")           # excluded from es
+    curator: bool = False                                               # excluded from es
+    dbPassword: str = Field(..., alias="password")                      # excluded from es
+    email: str = Field(..., alias="email")                              # excluded from es
     firstName: str = Field(None, alias="firstname")
+    fullName: str = Field(None)                                         # assigned as not_analyzed in es
+    joinDate: datetime.datetime = Field(None, alias="joindate")         # excluded from es
     lastName: str = Field(None, alias="lastname")
-    address: str = Field(None, alias="address")
-    affiliation: str = Field(None, alias="affiliation")
-    affiliationUrl: str = Field(None, alias="affiliationurl")
-    status: str = Field(..., alias="status")
-    role: str = Field(..., alias="role")
     orcid: str = Field(None, alias="orcid")
-    mobilePhoneNumber: str = None
-    officePhoneNumber: str = None
-    curator: bool = False
-    fullName: str = None
+    role: str = Field(..., alias="role")                                # excluded from es
+    status: str = Field(..., alias="status")                            # excluded from es
+    userId: int = Field(..., alias="id")                                # excluded from es
+    userName: str = Field(..., alias="username")                        # assigned as not_analyzed in es
+    userVerifyDbPassword: str = None                                    # not in es index mapping
+    mobilePhoneNumber: str = None                                       # not in es index mapping
+    officePhoneNumber: str = None                                       # not in es index mapping
+    apiToken: str = None                                                # excluded from es
 
     class Config:
         orm_mode = True
 
 
 class OrganismModel(BaseModel):
-    organismName: str = None
-    organismPart: str = None
+    organismName: str = None    # assigned as not_analyzed in es
+    organismPart: str = None    # assigned as not_analyzed in es
 
 
 class EntityModel(BaseModel):
@@ -41,29 +42,29 @@ class EntityModel(BaseModel):
 
 
 class ValidationEntryModel(BaseModel):
-    id: int
+    id: int = 0
     description: str = None
-    status: str
+    status: str = None
     statusExt: str = None
     group: str = None
-    passedRequirement: bool
+    passedRequirement: bool = None
     type: str = None
     message: str = None
-    overriden: bool
+    overriden: bool = False
 
 
 class ValidationEntriesModel(BaseModel):
     entries: List[ValidationEntryModel] = []
-    status: str
-    passedMinimumRequirement: bool
-    overriden: bool
+    status: str = None
+    passedMinimumRequirement: bool = None
+    overriden: bool = False
 
     class Config:
         orm_mode = True
 
 
 class StudyFactorModel(BaseModel):
-    name: str = None
+    name: str = None    # assigned as not_analyzed in es
 
 
 class FieldModel(BaseModel):
@@ -80,17 +81,17 @@ class TableModel(BaseModel):
 
 
 class BackupModel(BaseModel):
-    backupTimeStamp: datetime.datetime
-    backupId: str
-    folderPath: str
+    backupTimeStamp: int = None
+    backupId: str = None
+    folderPath: str = None
 
 
 class StudyDesignDescriptor(BaseModel):
-    description: str = None
+    description: str = None     # assigned as not_analyzed in es
 
 
 class PublicationModel(BaseModel):
-    abstractText: str = None
+    abstractText: str = None    # not in es index mapping
     title: str = None
     doi: str = None
     pubmedId: str = None
@@ -99,7 +100,7 @@ class PublicationModel(BaseModel):
 
 class ProtocolModel(BaseModel):
     name: str = None
-    Description: str = None
+    description: str = None
 
 
 class SampleMeasurementModel(BaseModel):
@@ -109,31 +110,31 @@ class SampleMeasurementModel(BaseModel):
 
 class MetaboliteAssignmentLine(BaseModel):
     identifier: str = None
-    databaseIdentifier: str = None
+    databaseIdentifier: str = None                          # in es mapping
     unitId: str = None # mzTab internal identificator
-    chemicalFormula: str = None
-    smiles: str  = None # V2 field only
-    inchi: str  = None # V2 field only
-    metaboliteIdentification: str  = None # V2 field
+    chemicalFormula: str = None                             # in es mapping
+    smiles: str  = None # V2 field only                     # in es mapping
+    inchi: str  = None # V2 field only                      # in es mapping
+    metaboliteIdentification: str  = None # V2 field        # in es mapping
     chemicalShift: str = None
     multiplicity: str = None
-    massToCharge: str = None
+    massToCharge: str = None                                # in es mapping
     fragmentation: str = None
     modifications: str = None  # V2 field only
     charge: str = None
-    retentionTime: str = None
-    taxid: str = None
-    species: str = None
+    retentionTime: str = None                               # in es mapping
+    taxid: str = None                                       # in es mapping
+    species: str = None                                     # in es mapping
     database: str = None
     databaseVersion: str = None
-    reliability: str = None
+    reliability: str = None                                 # in es mapping
     uri: str = None
     searchEngine: str = None
     searchEngineScore: str = None
     smallmoleculeAbundanceSub: str = None
     smallmoleculeAbundanceStdevSub: str = None
     smallmoleculeAbundanceStdErrorSub: str = None
-    sampleMeasurements: List[SampleMeasurementModel] = []
+    sampleMeasurements: List[SampleMeasurementModel] = []   # in es mapping
     assayName: str = None# This is the name of the Assay record this MAF is assigned to
 
 
@@ -143,13 +144,13 @@ class MetaboliteAssignmentModel(BaseModel):
 
 
 class AssayModel(BaseModel):
-    measurement: str = None
-    technology: str = None
+    measurement: str = None                                 # assigned as not_analyzed
+    technology: str = None                                  # assigned as not_analyzed
     platform: str = None
-    fileName: str = None
-    assayNumber: int = -1
-    metaboliteAssignment: MetaboliteAssignmentModel = None
-    assayTable: TableModel = None
+    fileName: str = None                                    # excluded from es
+    assayNumber: int = -1                                   # excluded from es
+    metaboliteAssignment: MetaboliteAssignmentModel = None  # excluded from es
+    assayTable: TableModel = None                           # excluded from es
 
 
 class ContactModel(BaseModel):
@@ -172,15 +173,15 @@ class StudySummaryModel(BaseModel):
 
 class LiteStudyModel(EntityModel):
     id: int = Field(...)
-    studyIdentifier: str = Field(...)
+    studyIdentifier: str = Field(...)           # assigned as not_analyzed in es
     title: Optional[str] = None
 
     studyDescription: str = None
-    studyStatus: str = None
+    studyStatus: str = None                     # assigned as not_analyzed in es
     studyPublicReleaseDate: int = Field(0)
     updateDate: int = Field(0)
     studySubmissionDate: int = Field(0)
-    obfuscationCode: str = Field(...)
+    obfuscationCode: str = Field(...)           # assigned as not_analyzed in es
     studySize: int = Field(0)
     validations: ValidationEntriesModel = None
     factors: List[StudyFactorModel] = []
@@ -195,15 +196,16 @@ class LiteStudyModel(EntityModel):
 
 
 class StudyModel(LiteStudyModel):
+    ObjectType: str = "Study"
     description: Optional[str]
-    studyLocation: Optional[str]
+    studyLocation: Optional[str]                            # excluded from es
     descriptors: List[StudyDesignDescriptor] = []
     publications: List[PublicationModel] = []
     protocols: List[ProtocolModel] = []
     assays: List[AssayModel] = []
-    contacts: List[ContactModel] = []
+    contacts: List[ContactModel] = []                       # excluded from es
     backups: List[BackupModel] = []
-    sampleTable: TableModel = None
+    sampleTable: TableModel = None                          # excluded from es
 
     @validator('updateDate', 'studySubmissionDate', 'studyPublicReleaseDate')
     def datetime_validation(cls, value):
