+++ app/ws/reports.py
@@ -30,7 +30,7 @@ from app.ws.isaApiClient import IsaApiClient
 from app.ws.misc_utilities.request_parsers import RequestParsers
 from app.ws.mtblsWSclient import WsClient
 from app.ws.ontology_info import *
-from app.ws.study_files import get_all_files
+from app.ws.study.folder_utils import get_all_files
 from app.ws.utils import log_request, writeDataToFile, readDatafromFile, clean_json, get_techniques, get_studytype, \
     get_instruments_organism
 from app.ws.report_builders.europe_pmc_builder import EuropePmcReportBuilder
@@ -674,7 +674,7 @@ class CrossReferencePublicationInformation(Resource):
         priv_list = wsc.get_private_studies()['content']
 
         msg = EuropePmcReportBuilder(priv_list, user_token, wsc, iac).build(drive)
-        if msg.count('Problem') is 1:
+        if msg.count('Problem') == 1:
             abort(500, msg)
 
         return 200, msg
