+++ requirements.txt
@@ -37,8 +37,9 @@ pyopenms==2.4.0
 python-Levenshtein==0.12.2
 plotly==4.6.0
 pillow==7.1.1
-requests==2.23.0
+requests==2.24.0
 recommonmark==0.6.0
 xmltodict==0.12.0
 zeep==3.4.0
 pydantic[dotenv]==1.9.0
+Elasticsearch==1.4.0
