+++ tests/mapStudies.py
@@ -30,17 +30,18 @@ import pandas as pd
 from flask import current_app as app
 from owlready2 import urllib
 
+from app.ws.study.study_service import StudyService
+
 
 class getStudyInfo():
 
     def __init__(self, studyID):
         try:
-            url = 'https://www.ebi.ac.uk/metabolights/webservice/study/' + studyID
-            request = urllib.request.Request(url)
-            request.add_header('user_token', 'b6cb38b7-8504-43bf-9281-a0c68fc06263')
-            response = urllib.request.urlopen(request)
-            content = response.read().decode('utf-8')
-            self.study_content = json.loads(content)
+            m_study = StudyService.get_instance().get_study_from_db_and_folder(studyID, user_token,
+                                                                               optimize_for_es_indexing=False,
+                                                                               revalidate_study=True,
+                                                                               include_maf_files=False)
+            self.study_content = {"content": m_study.dict()}
         except:
             print('cant find study', studyID)
 
@@ -79,13 +80,9 @@ class getStudyInfo():
 
 
 def searchStudies(query, feature='factors'):
-    url = 'https://www.ebi.ac.uk/metabolights/webservice/study/list'
-    request = urllib.request.Request(url)
-    request.add_header('user_token', 'b6cb38b7-8504-43bf-9281-a0c68fc06263')
-    response = urllib.request.urlopen(request)
-    content = response.read().decode('utf-8')
-    j_content = json.loads(content)
-
+    user_token = "b6cb38b7-8504-43bf-9281-a0c68fc06263"
+    study_id_list = StudyService.get_instance().get_all_authorized_study_ids(user_token)
+    j_content ={'content': study_id_list}
     import re
 
     def atoi(text):
